requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
qrcode[pil]==7.4.1
python-dotenv==1.0.0